dom/tests/mochitest/bugs/test_bug665548.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=665548
-->
<head>
  <title>Test for Bug 665548</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=665548">Mozilla Bug 665548</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name='location'></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 665548 **/

isnot(window.location, window.frames[0],
      "window.location shouldn't be the frame named location");

</script>
</pre>
</body>
</html>
